+++ src/components/display/table/index.ts
@@ -1,0 +1,8 @@
+export {default as Container} from "./TableContainer.svelte";
+
+export {default as Column} from "./TableColumn.svelte";
+export {default as Footer} from "./TableFooter.svelte";
+export {default as Header} from "./TableHeader.svelte";
+export {default as Heading} from "./TableHeading.svelte";
+export {default as Row} from "./TableRow.svelte";
+export {default as Section} from "./TableSection.svelte";