+++ src/components/interactables/check/stories/CheckDefaultStory.svelte
@@ -1,0 +1,6 @@
+<script>
+    import Check from "../Check.svelte";
+
+</script>
+
+<Check />