+++ src/components/interactables/form/FormHelpText.svelte
@@ -1,0 +1,33 @@
+<script lang="ts">
+    // TODO: Stories (?)
+
+    import type {DESIGN_HIDDEN_ARGUMENT} from "../../../lib/types/hidden";
+    import type {DESIGN_SPACING_ARGUMENT} from "../../../lib/types/spacings";
+
+    export let element: HTMLElement | null = null;
+
+    let _class: string = "";
+    export let id: string = "";
+    export let name: string = "";
+    export let style: string = "";
+    export let tabindex: number | string = "";
+    export let title: string = "";
+
+    export {_class as class};
+
+    export let hidden: DESIGN_HIDDEN_ARGUMENT = false;
+
+    export let margin: DESIGN_SPACING_ARGUMENT | undefined = undefined;
+
+    export let margin_x: DESIGN_SPACING_ARGUMENT | undefined = undefined;
+    export let margin_y: DESIGN_SPACING_ARGUMENT | undefined = undefined;
+
+    export let margin_top: DESIGN_SPACING_ARGUMENT | undefined = undefined;
+    export let margin_left: DESIGN_SPACING_ARGUMENT | undefined = undefined;
+    export let margin_bottom: DESIGN_SPACING_ARGUMENT | undefined = undefined;
+    export let margin_right: DESIGN_SPACING_ARGUMENT | undefined = undefined;
+</script>
+
+<small bind:this={element} {...$$props} class="form-help-text {_class}">
+    <slot />
+</small>