+++ src/components/interactables/radio/stories/RadioDefaultStory.svelte
@@ -1,0 +1,6 @@
+<script>
+    import Radio from "../Radio.svelte";
+
+</script>
+
+<Radio />