+++ src/components/interactables/switch/stories/SwitchDefaultStory.svelte
@@ -1,0 +1,6 @@
+<script>
+    import Switch from "../Switch.svelte";
+
+</script>
+
+<Switch />