+++ src/components/interactables/textinput/stories/TextInputTextAreaStory.svelte
@@ -1,0 +1,5 @@
+<script>
+    import TextInput from "../TextInput.svelte";
+</script>
+
+<TextInput {...$$props} is="textarea" value="This is a TextArea TextInput" />