+++ src/components/typography/heading/stories/heading.stories.js
@@ -18,7 +18,7 @@ export default {
     title: "Typography/Heading",
     component: Heading,
     argTypes: {
-        as: {
+        is: {
             options: ["h1", "h2", "h3", "h4", "h5", "h6"],
             control: {
                 type: "select",