+++ src/lib/types/variations.ts
@@ -1,21 +1,54 @@
 /**
- * Represents the tiers of variation tokens that can be applied to Framework Components
+ * Represents the tiers of filling variation tokens that can be applied to Framework Components
  */
-export enum DESIGN_VARIATION {
+export enum DESIGN_FILL_VARIATION {
     block = "block",
 
     clear = "clear",
 
+    flush = "flush",
+
     outline = "outline",
 }
 
-export const DESIGN_VARIATION_LITERALS = {...DESIGN_VARIATION} as const;
+/**
+ * Represents the tiers of flex variations tokens that can be applied to Framework Components
+ */
+export enum DESIGN_FLEX_VARIATION {
+    wrap = "wrap",
+}
 
-export type DESIGN_VARIATION_ARGUMENT = keyof typeof DESIGN_VARIATION_LITERALS;
+/**
+ * Represents the tiers of table variations tokens that can be applied to Framework Components
+ */
+export enum DESIGN_TABLE_VARIATION {
+    borders = "borders",
 
-export type DESIGN_VARIATION_CONTENT_ARGUMENT = Exclude<
-    DESIGN_VARIATION_ARGUMENT,
-    "block" | "clear"
+    stripes = "stripes",
+}
+
+export const DESIGN_FILL_VARIATION_LITERALS = {...DESIGN_FILL_VARIATION} as const;
+
+export const DESIGN_FLEX_VARIATION_LITERALS = {...DESIGN_FLEX_VARIATION} as const;
+
+export type DESIGN_FILL_VARIATION_ARGUMENT = keyof typeof DESIGN_FILL_VARIATION_LITERALS;
+
+export type DESIGN_FILL_BUTTON_VARIATION_ARGUMENT = Exclude<
+    DESIGN_FILL_VARIATION_ARGUMENT,
+    "block" | "flush"
 >;
 
-export type DESIGN_VARIATION_INTERACTIVE_ARGUMENT = Exclude<DESIGN_VARIATION_ARGUMENT, "block">;
+export type DESIGN_FILL_INPUT_VARIATION_ARGUMENT = Exclude<
+    DESIGN_FILL_VARIATION_ARGUMENT,
+    "clear" | "outline"
+>;
+
+export const DESIGN_TABLE_VARIATION_LITERALS = {...DESIGN_TABLE_VARIATION} as const;
+
+export type DESIGN_FLEX_VARIATION_ARGUMENT =
+    | keyof typeof DESIGN_FLEX_VARIATION_LITERALS
+    | (keyof typeof DESIGN_FLEX_VARIATION_LITERALS)[];
+
+export type DESIGN_TABLE_VARIATION_ARGUMENT =
+    | keyof typeof DESIGN_TABLE_VARIATION_LITERALS
+    | (keyof typeof DESIGN_TABLE_VARIATION_LITERALS)[];